refactor: add sorting flag
I don't expect any other sorting types will be added
so the logic consist of checking whether it is
`newest` or not.

One could argue with why wouldn't I make this a boolean.
My answer would be, in my opinion, it lose its meaning
because this is not something we want to enable or disable
but something that we want to decide which pattern
we want to use. So it is more like a semantic choice.

Signed-off-by: Taylan Dogan <git@taylandogan.info>