fix(ci): do not list maturin as a build dependency
Some platforms fail `pip install` because they don't have maturin available on pypi and maturin is listed as a build dependency. As far as I understand it, configuring the build-system is only needed if we want to let users build the wheel from source when they run `pip install`. Since we provide pre-built wheels I think this isn't that useful..

A side effect of this change is that the maturin action will automatically use the latest version of maturin unless otherwise configured via one of its inputs.