chore(examples): improve Github templates
The `New contributors` section should be within (under) the release
section, but they were on the same level (`##`).
This change adjusts the heading level to `###`.

There was no empty line between the list of new contributors and the
next release.
There should also be an empty line before the footer.
This change takes care of both of these cases.

Strip the v from the version number.

use ... in compare links